online_chat/requirements.txt
altgraph==0.17.3
asgiref==3.7.2
auto-py-to-exe==2.37.0
bottle==0.12.25
bottle-websocket==0.2.9
cachetools==5.3.1
cffi==1.15.1
customtkinter==5.2.0
darkdetect==0.8.0
Django==4.2.4
django-filter==23.2
django-jalali==6.0.1
django-jalali-date==1.0.2
django-ranged-response==0.2.0
django-recaptcha==3.0.0
django-simple-captcha==0.5.18
djangorestframework==3.14.0
Eel==0.16.0
et-xmlfile==1.1.0
future==0.18.3
gevent==23.7.0
gevent-websocket==0.10.1
greenlet==2.0.2
jdatetime==4.1.1
mysqlclient==2.2.0
numpy==1.25.2
openpyxl==3.1.2
pefile==2023.2.7
Pillow==10.0.0
psycopg2==2.9.7
pycparser==2.21
pyglet==2.0.9
pyinstaller==5.13.2
pyinstaller-hooks-contrib==2023.8
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
pywin32-ctypes==0.2.2
six==1.16.0
sqlparse==0.4.4
ttkbootstrap==1.10.1
tzdata==2023.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==6.0